Add MessagesTable.jsx under client Components
