+++ client/src/App.js
@@ -1,5 +1,7 @@
-import { Fragment, useEffect, useState } from 'react';
+import { useEffect, useState } from 'react';
 import logo from './logo.svg';
+// Importing the new component MessageTable
+import MessageTable from "./Components/MessagesTable";
 import './App.css';
 import axios from 'axios';
 
@@ -7,13 +9,6 @@ import Box from '@mui/material/Box';
 import Grid from '@mui/material/Unstable_Grid2';
 import TextField from '@mui/material/TextField';
 import Button from '@mui/material/Button';
-
-import Table from '@mui/material/Table';
-import TableBody from '@mui/material/TableBody';
-import TableCell from '@mui/material/TableCell';
-import TableContainer from '@mui/material/TableContainer';
-import TableHead from '@mui/material/TableHead';
-import TableRow from '@mui/material/TableRow';
 
 
 // rendering the table, the components to send a text message 
@@ -27,16 +22,6 @@ function App() {
   const[successfullResponse, setSuccessfullResponse] =useState(null);
   const[sentMessage,setSentMessage]= useState([]);
 
-  // IF don't put empty array, get call everytime 
-  useEffect(()=>{
-    // console.log("CALLING USEEFFECT:::");
-    axios.get("http://localhost:3000/customer-texts")
-    .then((response)=>{
-      // console.log(response.data);
-      setSentMessage(response.data);
-    })
-    .catch((e)=> {console.log(e);})
-  }, []); //from response into, use setSentMessage
 
   const retrieveMessages = async()=>{
     // console.log("CALLING USEEFFECT:::");
@@ -87,8 +72,6 @@ function App() {
   };
 
   // get all the messages
-
-
   //iteration over the array
   const renderMessage= () =>{
     return sentMessage.map((message, index) =>{
@@ -106,22 +89,6 @@ function App() {
       )
     });
   };
-
-  //do the same thing[const unas, or function]
-  //function submit() {}
-  const renderTableRows=()=>{
-    return sentMessage.map((message) =>{
-    return(
-      <TableRow key={message.phoneNumber+message.message}>
-      <TableCell>{message.phoneNumber}</TableCell>
-      <TableCell align='right'>{message.message}</TableCell>
-      <TableCell align='right'>{message.createdAt}</TableCell>
-      </TableRow>
-    );
-    });
-  };
-
-
 
   return (
     <div className="App">
@@ -160,18 +127,8 @@ function App() {
             </Grid>
           </Grid>
           <Grid xs={7}>
-          <TableContainer>
-            <Table sx={{ minWidth: 650 }} aria-label="simple table">
-              <TableHead>
-                <TableRow>
-                  <TableCell>Name</TableCell>
-                  <TableCell align="right">Phone Number</TableCell>
-                  <TableCell align="right">Date Sent</TableCell>
-                </TableRow>
-              </TableHead>
-              <TableBody>{renderTableRows()}</TableBody>
-            </Table>
-          </TableContainer>
+          <MessageTable/>
+{/* not passing any children */}
           </Grid>
         </Grid>
       </Box>
